+++ joulescope_ui/widgets/__init__.py
@@ -16,9 +16,9 @@ from .double_slider import DoubleSlider
 from .draggable_list_widget import DraggableListWidget
 
 from .accumulator import AccumulatorWidget
+from .clock import ClockWidget
 from .device_control import DeviceControlWidget
 from .debug import DebugWidget
-from .publish_spy import PublishSpyWidget
 # from .example import ExampleWidget
 from .memory import MemoryWidget
 from .hamburger import HamburgerWidget
@@ -27,6 +27,7 @@ from .jls_info import JlsInfoWidget
 from .js220_cal import JS220CalibrationWidget
 from .notes import NotesWidget
 from .progress_bar import ProgressBarWidget
+from .publish_spy import PublishSpyWidget
 from .record_status import RecordStatusWidget
 from .sidebar import SideBar
 from .settings import SettingsWidget