+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.0.49
 
-2024 Jan 26 [in progress]
+2024 Jan 30 [in progress]
 
 * Dropped Python 3.9 support as static methods are not callable until 3.10.
   See https://docs.python.org/3/whatsnew/3.10.html
@@ -27,6 +27,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Added time format options to CSV statistics recording  #231
 * Added option to set exact plot y-axis range in Waveform widget  #126
 * Added support for Waveform widget precision and quantity selection  #130
+* Added clock widget with support for local and UTC time.
 
 
 ## 1.0.48